tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-timing-function_inline.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-timing-function and inline element</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-timing-function_tag">
   <meta name="flags" content="dom">
   <meta name="assert" content="animation-timing-function applies to inline
   elements">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes fade {
         from {
            opacity: 1;
         }
         to {
            opacity: 0;
         }
      }
      .test {
         animation-name: fade;
         animation-duration: 1s;
         animation-fill-mode: forwards;
         animation-timing-function: step-end;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var span = document.querySelector('span');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(span, null);
         var passed = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         setTimeout(function() {
            passed = (style.opacity == 1);
         }, 750);
         span.addEventListener('animationend', function() {
            clearTimeout(timeout);
            passed = (passed && style.opacity == 0);
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         span.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The uppercase word below should abruptly disappear after one second.</p>
   <p>Testing...</p>
   <div>normal <span>DISAPPEAR</span> normal</div>
</body>
</html>
